# src/main/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Control API
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Local Dev
  - url: http://localhost:8080
    description: Docker
paths:
  /api/users:
    get:
      tags: [User]
      summary: Liste aller User
      operationId: listUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [User]
      summary: User anlegen
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /api/user/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [User]
      summary: User per ID
      operationId: getUserById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Nicht gefunden
    put:
      tags: [ User ]
      summary: User aktualisieren
      operationId: updateUserById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Nicht gefunden
    delete:
      tags: [User]
      summary: User löschen
      operationId: deleteUserById
      responses:
        '204':
          description: Gelöscht (kein Inhalt)
        '404':
          description: No content





components:
  schemas:
    User:
      type: object
      properties:
        id:   { type: integer, format: int64, example: 42 }
        name: { type: string, example: Alice }
      required: [id, name]

    CreateUserRequest:
      type: object
      properties:
        name: { type: string }
      required: [name]